Added packets for Lepton VOSPI and worked out some bugs in the library.
It was the first time that the STM32F4 was sending packets using the
utility and the PC side was receiving packets using the utility.  So
some bugs were found.

Still need to work out the VOSPI CRC.  It seems that a decent number
of packets are failing.  I need to determine if it is due to bad CRC
algorithm, or if it is becuase the SPI bus is set up wrong or has
some timing issues.