Any size circular buffer and alignment padding.
Instead of having a hard coded circular buffer size, the user can
now provide a pointer to a GenericPacket array and the size of that
array.  More dangerous...but also more flexible.  Helpfil since
this code runs on both uC and PC hardware.

A feature for alignment padding was added.  This is because a lot
of the uC DMA operations I am using will not flush the DMA buffer
until all 4 bytes are in the FIFO.  If you have a packet that
doesn't fall on that boundary...the last few bytes could be stuck.
By shorteninng the usable size of the GenericPacket, I can
guarantee that there are always at least 4 more bytes at the end of
the array that I can push out to force the DMA FIFO boundary to
trip.